backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
asyncpg>=0.27.0
python-dotenv>=1.0.0
alembic>=1.10.0
pydantic>=2.0.0
